+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### New Features
 
 - Added support for `Series.str.ljust` and `Series.str.rjust`.
+- Added support for `Series.str.center`.
 
 
 ## 1.26.0 (2024-12-05)